refactor(rx32-alterando-repositorio-categoria): corrige e melhora repositorios de categoria
