+++ src/modules/cars/infra/typeorm/repositories/SpecificationsRepository.ts
@@ -1,15 +1,17 @@
+import { getRepository, Repository } from "typeorm";
 
 import {
-  ICreateSpecificationDTO, 
+  ICreateSpecificationDTO,
   ISpecificationsRepository,
-} from "../ISpecificationRepository";
+} from "@modules/cars/repositories/ISpecificationsRepository";
 
-import { Specification } from "../../model/Specification";
+import { Specification } from "../entities/Specification";
 
 class SpecificationsRepository implements ISpecificationsRepository {
   private repository: Repository<Specification>;
 
   constructor() {
+    this.repository = getRepository(Specification);
   }
 
   async create({
