+++ src/modules/cars/repositories/ISpecificationsRepository.ts
@@ -1,4 +1,4 @@
-import { Specification } from "../model/Specification";
+import { Specification } from "@modules/cars/infra/typeorm/entities/Specification";
 
 interface ICreateSpecificationDTO {
   name: string;
