+++ src/modules/cars/useCases/createSpecification/CreateSpecificationUseCase.ts
@@ -1,4 +1,4 @@
-import { ISpecificationsRepository } from "../../repositories/ISpecificationRepository"
+import { ISpecificationsRepository } from "../../repositories/ISpecificationsRepository"
 
 interface IRequest{
   name: string;
